cleanup layer_views, move LayerMap generator to its own module
