+++ gdsfactory/cli.py
@@ -12,7 +12,7 @@ import gdsfactory
 from gdsfactory.config import cwd, print_config
 from gdsfactory.generic_tech import LAYER
 from gdsfactory.install import install_gdsdiff, install_klayout_package
-from gdsfactory.technology.layer_views import lyp_to_dataclass
+from gdsfactory.technology import lyp_to_dataclass
 from gdsfactory.types import PathType
 from gdsfactory.write_cells import write_cells as write_cells_to_separate_gds
