+++ gdsfactory/technology/__init__.py
@@ -1,3 +1,4 @@
+from gdsfactory.technology.layer_map import lyp_to_dataclass
 from gdsfactory.technology.layer_stack import LayerLevel, LayerStack
 from gdsfactory.technology.layer_views import LayerView, LayerViews
 
@@ -6,4 +7,5 @@ __all__ = [
     "LayerViews",
     "LayerLevel",
     "LayerStack",
+    "lyp_to_dataclass",
 ]
